Add simple events in the client provided by the lecturers.
This commit adds hooks into the client so that we can register
for events in the GUI code without rewriting the entire client.

It also implements the required functions in ChatWindowController
to handle message receive events, room change events, client
joined room events, client left room events.

It also adds a current room client list to the left hand side of
the GUI.